+++ ex01/src/main.cpp
@@ -33,7 +33,7 @@ int main()
 		span3.addMultiple(55, 55);
 		int arr[] = {1, 6, 5, 8, 9, 7, 8, 1, 3, 2};
 		std::vector<int> vec(arr, arr + sizeof(arr)/sizeof(arr[0]));
-		span3.addMultiple(vec.size(), vec);
+		span3.addMultiple(static_cast<unsigned int>(vec.size()), vec);
 		span3.print();
 	}
 	catch(const std::exception& e)
